http://download.virtualbox.org/virtualbox/4.2.0_BETA1/VirtualBox-4.2.0_BETA1.tar.bz2

Merge commit '107efe57f105c40e1f7e9448e2f6e5f2001cd954' of git://github.com/larsr/VirtualBox-OSE into VirtualBox/4.2/tar

Conflicts:
	include/VBox/com/VirtualBoxErrorInfo.h
	include/iprt/darwin/machkernel.h
	src/VBox/Additions/WINNT/Graphics/Video/disp/wddm/VBoxDispD3DCmn.h
	src/VBox/Additions/WINNT/Graphics/Video/mp/common/VBoxMPCommon.h
	src/VBox/Additions/WINNT/Graphics/Video/mp/common/VBoxMPUtils.h
	src/VBox/Additions/WINNT/Graphics/Video/mp/wddm/VBoxMPVdma.h
	src/VBox/Additions/WINNT/Graphics/Video/mp/wddm/VBoxMPWddm.cpp
	src/VBox/Additions/WINNT/Graphics/Video/mp/wddm/VBoxMPWddm.h
	src/VBox/Additions/WINNT/Installer/VBoxGuestAdditions.nsi
	src/VBox/Additions/WINNT/Mouse/NT5/VBoxMF.h
	src/VBox/Additions/WINNT/Mouse/NT5/VBoxMFDriver.cpp
	src/VBox/Additions/WINNT/Mouse/common/VBoxMouseLog.h
	src/VBox/Additions/WINNT/SharedFolders/redirector/Makefile.kmk
	src/VBox/Additions/WINNT/SharedFolders/redirector/dll/Makefile.kmk
	src/VBox/Additions/WINNT/SharedFolders/redirector/dll/vboxmrxp.rc
	src/VBox/Additions/WINNT/SharedFolders/redirector/sys/Makefile.kmk
	src/VBox/Additions/WINNT/SharedFolders/redirector/sys/VBoxSF.rc
	src/VBox/Additions/WINNT/VBoxCredProv/VBoxCredPoller.cpp
	src/VBox/Additions/WINNT/VBoxCredProv/VBoxCredPoller.h
	src/VBox/Additions/WINNT/VBoxCredProv/VBoxCredProv.rc
	src/VBox/Additions/WINNT/i8042prt/Makefile.kmk
	src/VBox/Additions/WINNT/i8042prt/i8042prt.rc
	src/VBox/Additions/common/VBoxGuestLib/VBGLInternal.h
	src/VBox/Additions/common/VBoxService/VBoxServiceControlExec.cpp
	src/VBox/Additions/common/VBoxService/VBoxServiceControlExecThread.cpp
	src/VBox/Additions/common/VBoxService/VBoxServiceControlExecThread.h
	src/VBox/Additions/common/VBoxService/VBoxServicePipeBuf.cpp
	src/VBox/Additions/common/VBoxService/VBoxServicePipeBuf.h
	src/VBox/Additions/common/pam/pam_vbox.c
	src/VBox/Additions/linux/installer/90-vboxguest.fdi
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxPkgOSE.dsc
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxPkgOSE.fdf
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxPkgOSEX64.dsc
	src/VBox/Devices/Network/solaris/vbox-libdlpi.cpp
	src/VBox/Devices/Network/solaris/vbox-libdlpi.h
	src/VBox/Devices/Storage/VSCSI/VSCSISgBuf.cpp
	src/VBox/Disassembler/DisasmTestA.asm
	src/VBox/Frontends/VBoxBalloonCtrl/VBoxBalloonCtrl.cpp
	src/VBox/Frontends/VBoxBalloonCtrl/VBoxBalloonCtrl.h
	src/VBox/Frontends/VBoxHeadless/FramebufferVNC.cpp
	src/VBox/Frontends/VBoxHeadless/FramebufferVNC.h
	src/VBox/Frontends/VirtualBox/src/VBoxMediaManagerDlg.cpp
	src/VBox/Frontends/VirtualBox/src/VBoxMedium.cpp
	src/VBox/Frontends/VirtualBox/src/VBoxVMLogViewer.cpp
	src/VBox/Frontends/VirtualBox/src/extensions/QIWizard.cpp
	src/VBox/Frontends/VirtualBox/src/globals/VBoxDefs.cpp
	src/VBox/Frontends/VirtualBox/src/widgets/VBoxOSTypeSelectorWidget.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/clonevm/UICloneVMWizard.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/exportappliance/UIExportApplianceWzd.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/firstrun/UIFirstRunWzd.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/importappliance/UIImportApplianceWzd.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/newhd/UINewHDWizard.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/newvm/UINewVMWzd.cpp
	src/VBox/Frontends/VirtualBox/src/wizards/registration/UIRegistrationWzd.cpp
	src/VBox/HostDrivers/Support/SUPDrv-dtrace.d
	src/VBox/HostDrivers/Support/SUPDrv.c
	src/VBox/HostDrivers/Support/SUPR0.def
	src/VBox/HostDrivers/Support/linux/Makefile
	src/VBox/HostDrivers/Support/solaris/mod.sh
	src/VBox/HostDrivers/VBoxNetFlt/win/nobj/VBoxNetFltNobj.def
	src/VBox/Main/cbinding/VBoxXPCOMCGlue.h
	src/VBox/Main/glue/VirtualBoxErrorInfo.cpp
	src/VBox/Main/glue/tests/TestVBox.java
	src/VBox/Main/include/PciDeviceAttachmentImpl.h
	src/VBox/Main/include/PciRawDevImpl.h
	src/VBox/Main/src-all/Global.cpp
	src/VBox/Main/src-all/PciDeviceAttachmentImpl.cpp
	src/VBox/Main/src-client/GuestCtrlIO.cpp
	src/VBox/Main/src-client/PciRawDevImpl.cpp
	src/VBox/Main/src-server/Logging.cpp
	src/VBox/Main/src-server/SnapshotImpl.cpp
	src/VBox/Main/src-server/SystemPropertiesImpl.cpp
	src/VBox/Main/src-server/VirtualBoxImpl.cpp
	src/VBox/Main/src-server/darwin/OpenGLTestDarwin.cpp
	src/VBox/Main/src-server/generic/OpenGLTest.cpp
	src/VBox/Main/src-server/generic/OpenGLTestApp.cpp
	src/VBox/Runtime/common/path/RTPathHaveExt.cpp
	src/VBox/Runtime/common/path/RTPathHavePath.cpp
	src/VBox/Runtime/common/string/RTStrConvertHexBytes.cpp
	src/VBox/Runtime/r0drv/darwin/mach_kernel-r0drv-darwin.cpp
	src/VBox/Runtime/r0drv/solaris/vbi/RTMpPokeCpu-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/alloc-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/memobj-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/mp-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/mpnotification-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/process-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/thread-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/thread2-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/time-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/timer-r0drv-solaris.c
	src/VBox/Runtime/testcase/tstErrUnique.cpp
	src/VBox/Runtime/tools/RTLdrFlt.cpp
	src/VBox/VMM/VMMR3/EMRaw.cpp
	src/VBox/VMM/VMMR3/VBoxVMMDeps.cpp
	src/VBox/VMM/testcase/tstCompiler.cpp
	src/libs/libxml2-2.6.30/INSTALL
	src/libs/libxml2-2.6.30/Makefile.kmk
	src/libs/libxml2-2.6.30/README
	src/libs/libxml2-2.6.30/TODO
	src/libs/libxml2-2.6.30/bakefile/Bakefiles.bkgen
	src/libs/libxml2-2.6.30/python/TODO
	src/libs/libxml2-2.6.30/trio.c
	src/libs/libxml2-2.6.30/trio.h
	src/libs/libxml2-2.6.30/triodef.h
	src/libs/libxml2-2.6.30/trionan.c
	src/libs/libxml2-2.6.30/trionan.h
	src/libs/libxml2-2.6.30/triop.h
	src/libs/libxml2-2.6.30/triostr.c
	src/libs/libxml2-2.6.30/triostr.h
	src/libs/libxslt-1.1.22/FEATURES
	src/libs/libxslt-1.1.22/Makefile.kmk
	src/libs/libxslt-1.1.22/README
	src/libs/libxslt-1.1.22/doc/xsltproc.1
	src/libs/libxslt-1.1.22/doc/xsltproc.xml
	src/libs/libxslt-1.1.22/libxslt/extra.c
	src/libs/libxslt-1.1.22/libxslt/trio.h
	src/libs/libxslt-1.1.22/libxslt/triodef.h
	src/libs/zlib-1.2.1/Makefile.kmk

+++ src/VBox/Frontends/VBoxSDL/Framebuffer.cpp
@@ -465,8 +465,8 @@ STDMETHODIMP VBoxSDLFB::NotifyUpdate(ULONG x, ULONG y,
     event.user.code  = mScreenId;
     event.user.type  = SDL_USER_EVENT_UPDATERECT;
     // 16 bit is enough for coordinates
-    event.user.data1 = (void*)(x << 16 | y);
-    event.user.data2 = (void*)(w << 16 | h);
+    event.user.data1 = (void*)(uintptr_t)(x << 16 | y);
+    event.user.data2 = (void*)(uintptr_t)(w << 16 | h);
     PushNotifyUpdateEvent(&event);
 #else /* !VBOXSDL_WITH_X11 */
     update(x, y, w, h, true /* fGuestRelative */);
@@ -1183,9 +1183,9 @@ void VBoxSDLFB::paintSecureLabel(int x, int y, int w, int h, bool fForce)
          && !mSecureLabelText.isEmpty()
        )
     {
-        SDL_Color clrFg = {(mSecureLabelColorFG & 0x00FF0000) >> 16,
-                           (mSecureLabelColorFG & 0x0000FF00) >> 8,
-                           mSecureLabelColorFG & 0x000000FF, 0};
+        SDL_Color clrFg = {(uint8_t)((mSecureLabelColorFG & 0x00FF0000) >> 16),
+                           (uint8_t)((mSecureLabelColorFG & 0x0000FF00) >> 8),
+                           (uint8_t)( mSecureLabelColorFG & 0x000000FF      ), 0};
         SDL_Surface *sText = (pTTF_RenderUTF8_Blended != NULL)
                                  ? pTTF_RenderUTF8_Blended(mLabelFont, mSecureLabelText.c_str(), clrFg)
                                  : pTTF_RenderUTF8_Solid(mLabelFont, mSecureLabelText.c_str(), clrFg);
